Shift create packet data dict functions to plot_util.py
